重新寫算法 第一次紅燈
寫法有問題造成無窮迴圈

+++ RobertBookStore/Warehouse.cs
@@ -38,7 +38,7 @@ namespace RobertBookStore
         /// <summary>
         /// 總價
         /// </summary>
-        public double GrossPrice { get; set; }
+        public decimal GrossPrice { get; set; }
     }
 
 }
